Refactor and test ingest_station_description.
